feat: move utilities and data classes to commons (batch 13)
Move platform-independent types to commons module for KMP iOS reuse:

commonMain:
- ServerName, ServerType, DEFAULT_MEDIA_SERVERS (media server config)
- StringFeedState (sealed class for feed states)
- BookmarkType (enum)
- SupportedContent (URL/MIME type filtering)
- AdditiveComplexFeedFilter (extends FeedFilter from commons)

jvmAndroid:
- ZapFormatterNoDecimals (integer amount formatting)

Original locations retain re-exports (typealiases/delegates) for
backwards compatibility. Consumers in the same package updated
with explicit imports where needed.
